+++ Sources/Main/Util/Private/AnalyticsManager+ReferralAppTracking.swift
@@ -9,4 +9,18 @@ extension AnalyticsManager {
         process(RAnalyticsEvent(name: AnalyticsManager.Event.Name.pageVisit, parameters: nil))
         launchCollector.referralTracking = .none
     }
+
+    func tryToTrackReferralApp(with url: URL?, sourceApplication: String?) {
+        if let url = url {
+            launchCollector.origin = .external
+            trackReferralApp(url: url, sourceApplication: sourceApplication)
+        }
+    }
+
+    func tryToTrackReferralApp(with webpageURL: URL?) {
+        if let url = webpageURL {
+            launchCollector.origin = .external
+            trackReferralApp(url: url, sourceApplication: nil)
+        }
+    }
 }